Some Settings work.
Also, I just learned that everything gets compiled together into a giant blob that doesn't have a file system.  Which makes the whole "hot swapping of skins" thing kinda moot for now.  And also means that all the File.Exists () calls for internal files ... don't work.

So, I have to start a rewrite to properly load the sprites if they exist, and separately try to load textures if those are preferred.